+++ src/core/api/providers/requesty.ts
@@ -2,8 +2,8 @@ import { Anthropic } from "@anthropic-ai/sdk"
 import { ModelInfo, requestyDefaultModelId, requestyDefaultModelInfo } from "@shared/api"
 import { calculateApiCostOpenAI } from "@utils/cost"
 import OpenAI from "openai"
+import { toRequestyServiceStringUrl } from "@/shared/clients/requesty"
 import { fetch } from "@/shared/net"
-import { toRequestyServiceStringUrl } from "@/shared/providers/requesty"
 import { ApiHandler, CommonApiHandlerOptions } from "../index"
 import { withRetry } from "../retry"
 import { convertToOpenAiMessages } from "../transform/openai-format"
