+++ src/shared/providers/bedrock.json
