+++ src/shared/providers/vertex.json
